progress from 11/16
