+++ config.js
@@ -1,0 +1,3 @@
+var routesApiKey = "200227251-dff824783300bfdde3f3714b65feede7";
+var trailsApiKey = "200227251-52bbf40f14de65bb8de3f88b82402869";
+var OpenWeatherAPI = "94e068ccf7cb21e0c2f4fa4540420c33";
