--- index.html
@@ -1,90 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title>Social Distance</title>
-  <script src="https://unpkg.com/simple-statistics@5.0.0/dist/simple-statistics.min.js"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/d3/5.9.7/d3.min.js"></script>
-  <script src="https://cdn.plot.ly/plotly-latest.min.js"></script>
-  <style>
-    body {
-      background-color: unset;
-    }
-/* table {
-    border-collapse: collapse;
-    width: 100%;
-} */
-
-th, td {
-    text-align: left;
-    padding: 8px;
-}
-tbody>tr:nth-child(even){background-color: #f2f2f2 !important}
-tbody>tr:nth-child(odd){background-color: #fff !important}
-/* th {
-  background-color: #2f3439;
-  color: white;
-} */
-  </style>
-</head>
-
-<body>
-  <div class="container">
-    <div class="jumbotron">
-      <h1>
-        routes
-        <span id="report-date" style="float: right;">2019</span>
-      </h1>
-      <p></p>
-    </div>
-
-    <div class="row">
-      <div class="col-md-4">
-        <div class="panel panel-primary">
-          <div class="panel-body">
-            Company Profile
-          </div>
-        </div>
-        <div class="company">
-          Going the Social Distance!
-        </div>
-      </div>
-      <div class="col-md-8">
-        <div class="panel panel-default">
-          <div class="panel-body">
-            Stock Performance
-          </div>
-        </div>
-        <div id="plot"></div>
-      </div>
-    </div>
-
-    <div class="row">
-      <div class="col-md-12">
-        <table id="summary-table" class="table table-striped table-inverse">
-          <thead>
-            <tr>
-              <th>Date</th>
-              <th>Open</th>
-              <th>High</th>
-              <th>Low</th>
-              <th>Close</th>
-              <th>Volume</th>
-            </tr>
-          </thead>
-          <tbody></tbody>
-        </table>
-      </div>
-    </div>
-  </div>
-
-
-  <script src="jsTests/data.js"></script>
-  
-
-</body>
-
-</html>
