+++ loc_list.csv
@@ -1,0 +1,6 @@
+name,lat,lon
+Sand Rock,34.18,-85.817
+Jamestown,34.411,-85.574
+Palisades Park,33.989,-86.458
+Anchorage & South Central Ice Climbing,61.213,-149.867
+Denali National Park,63.065,-151.172
